Remove setter Use Task instead of Service Use CellValueFactory
