+++ src/main/java/org/jabref/gui/actions/CopyFilesResultItemViewModel.java
@@ -1,5 +1,7 @@
 package org.jabref.gui.actions;
 
+import java.nio.file.Path;
+
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -7,36 +9,27 @@ import javafx.beans.property.StringProperty;
 
 import de.jensd.fx.glyphs.materialdesignicons.MaterialDesignIcon;
 
-public class CopyFilesResultViewModel {
+public class CopyFilesResultItemViewModel {
 
     private final StringProperty file = new SimpleStringProperty("");
     private final ObjectProperty<MaterialDesignIcon> icon = new SimpleObjectProperty<>(MaterialDesignIcon.ALERT);
 
     private final StringProperty message = new SimpleStringProperty("");
 
-    public CopyFilesResultViewModel(String file, boolean success, String message) {
-        this.file.setValue(file);
+    public CopyFilesResultItemViewModel(Path file, boolean success, String message) {
+        this.file.setValue(file.toString());
         this.message.setValue(message);
         if (success) {
             this.icon.setValue(MaterialDesignIcon.CHECK);
         }
-
     }
 
     public StringProperty getFile() {
         return file;
     }
 
-    public void setFile(String file) {
-        this.file.setValue(file);
-    }
-
     public StringProperty getMessage() {
         return message;
-    }
-
-    public void setMessage(String message) {
-        this.message.setValue(message);
     }
 
     public ObjectProperty<MaterialDesignIcon> getIcon() {
@@ -45,6 +38,6 @@ public class CopyFilesResultViewModel {
 
     @Override
     public String toString() {
-        return "CopyFilesResultViewModel [file=" + file.get() + ", message=" + message.get() + "]";
+        return "CopyFilesResultItemViewModel [file=" + file.get() + ", message=" + message.get() + "]";
     }
 }
