+++ src/main/java/org/jabref/gui/actions/ExportLinkedFilesAction.java
@@ -10,11 +10,10 @@ import javax.swing.AbstractAction;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.concurrent.Service;
+import javafx.concurrent.Task;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.Dialog;
 import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.paint.Color;
@@ -26,6 +25,7 @@ import org.jabref.gui.DialogService;
 import org.jabref.gui.FXDialogService;
 import org.jabref.gui.util.DefaultTaskExecutor;
 import org.jabref.gui.util.DirectoryDialogConfiguration;
+import org.jabref.gui.util.ValueTableCellFactory;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
@@ -58,29 +58,28 @@ public class ExportLinkedFilesAction extends AbstractAction {
         exportPath.ifPresent(path -> {
             databaseContext = JabRefGUI.getMainFrame().getCurrentBasePanel().getDatabaseContext();
 
-            Service<List<CopyFilesResultViewModel>> exportService = new ExportLinkedFilesService(databaseContext, entries, path);
-            startServiceAndshowProgessDialog(exportService);
+            Task<List<CopyFilesResultItemViewModel>> exportTask = new ExportLinkedFilesTask(databaseContext, entries, path);
+            startServiceAndshowProgessDialog(exportTask);
         });
     }
 
-    private void startServiceAndshowProgessDialog(Service<List<CopyFilesResultViewModel>> exportService) {
+    private void startServiceAndshowProgessDialog(Task<List<CopyFilesResultItemViewModel>> exportService) {
         DefaultTaskExecutor.runInJavaFXThread(() -> {
             exportService.setOnSucceeded(value -> {
                 DefaultTaskExecutor.runInJavaFXThread(() -> showDialog(exportService.getValue()));
             });
-            exportService.start();
-            DialogService dlgService = new FXDialogService();
-            dlgService.showCanceableProgressDialogAndWait(exportService);
+            exportService.run();
+            dialogService.showCanceableProgressDialogAndWait(exportService);
 
         });
     }
 
-    private void showDialog(List<CopyFilesResultViewModel> data) {
+    private void showDialog(List<CopyFilesResultItemViewModel> data) {
         Dialog<ButtonType> dlg = new Dialog<>();
         dlg.setTitle(Localization.lang("Result"));
-        ObservableList<CopyFilesResultViewModel> tableData = FXCollections.observableArrayList(data);
+        ObservableList<CopyFilesResultItemViewModel> tableData = FXCollections.observableArrayList(data);
 
-        TableView<CopyFilesResultViewModel> tv = createTable();
+        TableView<CopyFilesResultItemViewModel> tv = createTable();
         ScrollPane sp = new ScrollPane();
         sp.setContent(tv);
 
@@ -95,44 +94,29 @@ public class ExportLinkedFilesAction extends AbstractAction {
         dlg.showAndWait();
     }
 
-    private static TableView<CopyFilesResultViewModel> createTable() {
-        TableView<CopyFilesResultViewModel> tableResult = new TableView<>();
+    private static TableView<CopyFilesResultItemViewModel> createTable() {
+        TableView<CopyFilesResultItemViewModel> tableResult = new TableView<>();
 
-        TableColumn<CopyFilesResultViewModel, String> colFile = new TableColumn<>(Localization.lang("File"));
-        TableColumn<CopyFilesResultViewModel, MaterialDesignIcon> colIcon = new TableColumn<>(Localization.lang("Status"));
-        TableColumn<CopyFilesResultViewModel, String> colMessage = new TableColumn<>(Localization.lang("Message"));
+        TableColumn<CopyFilesResultItemViewModel, String> colFile = new TableColumn<>(Localization.lang("File"));
+        TableColumn<CopyFilesResultItemViewModel, MaterialDesignIcon> colIcon = new TableColumn<>(Localization.lang("Status"));
+        TableColumn<CopyFilesResultItemViewModel, String> colMessage = new TableColumn<>(Localization.lang("Message"));
 
         colFile.setCellValueFactory(cellData -> cellData.getValue().getFile());
         colMessage.setCellValueFactory(cellData -> cellData.getValue().getMessage());
         colIcon.setCellValueFactory(cellData -> cellData.getValue().getIcon());
 
-        colIcon.setCellFactory(column -> {
-            return new TableCell<CopyFilesResultViewModel, MaterialDesignIcon>() {
+        colIcon.setCellFactory(new ValueTableCellFactory<CopyFilesResultItemViewModel, MaterialDesignIcon>().withGraphic(item -> {
 
-                @Override
-                protected void updateItem(MaterialDesignIcon item, boolean empty) {
-                    super.updateItem(item, empty);
+            Text icon = MaterialDesignIconFactory.get().createIcon(item);
+            if (item == MaterialDesignIcon.CHECK) {
+                icon.setFill(Color.GREEN);
+            }
+            if (item == MaterialDesignIcon.ALERT) {
+                icon.setFill(Color.RED);
+            }
+            return icon;
+        }));
 
-                    if ((item == null) || empty) {
-                        setText(null);
-                        setStyle("");
-                    } else {
-                        Text icon = MaterialDesignIconFactory.get().createIcon(item);
-
-                        //Green checkmark
-                        if (item == MaterialDesignIcon.CHECK) {
-                            icon.setFill(Color.GREEN);
-                        }
-                        //Red Alert symbol
-                        if (item == MaterialDesignIcon.ALERT) {
-                            icon.setFill(Color.RED);
-                        }
-                        setGraphic(icon);
-
-                    }
-                }
-            };
-        });
         tableResult.getColumns().add(colIcon);
         tableResult.getColumns().add(colMessage);
         tableResult.getColumns().add(colFile);
